Extract components from App
.. and move .css to its own folder
