+++ src/App.js
@@ -2,359 +2,11 @@ import React, { Component } from 'react';
 import {
   BrowserRouter as Router,
   Route,
-  Link,
 } from 'react-router-dom';
-import { apiUrl } from './config';
-import Victim from './Victim';
-import loadingGif from './images/loading.gif';
-import resolveIcon from './iconResolve';
-var httpBuildQuery = require('http-build-query');
-
-class HeroSection extends Component {
-  render() {
-    return (
-      <section className="hero is-small is-dark gray">
-        <div className="hero-body">
-          <div className="container">
-            <h1 className="title">
-              <Link to="/">
-                Open Perpetuum Killboard
-              </Link>
-            </h1>
-            <h2 className="subtitle">
-              v0.3
-            </h2>
-          </div>
-        </div>
-      </section>
-    );
-  }
-}
-
-class FooterSection extends Component {
-  render() {
-    return (
-      <footer className="footer gray">
-        <div className="container">
-          <div className="content has-text-centered">
-            <p>
-              <strong>Open Perpetuum Killboard</strong> is part of Open Perpetuum, an open-source project. Support us on <a href="https://www.patreon.com/openperpetuum">Patreon</a>!
-            </p>
-          </div>
-        </div>
-      </footer>
-    );
-  }
-}
-
-class KillBoard extends Component {
-  constructor(props) {
-    super(props);
-    this.handleClickNext = this.handleClickNext.bind(this);
-    this.handleClickPrev = this.handleClickPrev.bind(this);
-    this.handleClickFirst = this.handleClickFirst.bind(this);
-    this.handleClickLast = this.handleClickLast.bind(this);
-    this.populateTableFromAPI = this.populateTableFromAPI.bind(this);
-    this.state = {
-      'kills': [],
-      'page': this.state != null ? this.state.page : 1,
-      'prev': this.state != null ? this.state.page : 1,
-      'page_count': this.state != null ? this.state.page_count : 1,
-      'loaded': false
-    };
-  }
-
-  handleClickNext() {
-    this.populateTableFromAPI(this.state.next);
-  }
-
-  handleClickPrev() {
-    this.populateTableFromAPI(this.state.prev);
-  }
-
-  handleClickFirst() {
-    this.populateTableFromAPI(1);
-  }
-
-  handleClickLast() {
-    this.populateTableFromAPI(this.state.page_count);
-  }
-
-
-  populateTableFromAPI(pageNum) {
-    this.setState({ 'loaded': false });
-    var searchParams = {
-      'order-by': [
-        {
-          'type': 'field',
-          'field': 'date',
-          'direction': 'desc',
-        },
-      ],
-      'page': pageNum
-    };
-
-    fetch(new URL('/killboard/kill?' + httpBuildQuery(searchParams), apiUrl), {
-      method: 'GET',
-      headers: new Headers({
-        "Accept": "application/*+json",
-      }),
-    })
-      .then(response => {
-        return response.json();
-      })
-      .then(json => {
-        this.setState({
-          'loaded': true,
-          'page': json.page,
-          'next': json.page + 1,
-          'prev': json.page - 1,
-          'page_count': json.page_count,
-          'kills': json._embedded.kill.map(function (kill) {
-            return (
-              <Victim
-                key={kill.id}
-                killId={kill.id}
-                victimRobot={kill._embedded.robot.name}
-                victimAgent={kill._embedded.agent.name}
-                victimCorporation={kill._embedded.corporation.name}
-                attackerAgent={
-                  kill._embedded.attackers.map(function (attacker) {
-                    if (attacker.hasKillingBlow === true) {
-                      return attacker._embedded.agent.name;
-                    }
-                    return '';
-                  })
-                }
-                attackerCorporation={
-                  kill._embedded.attackers.map(function (attacker) {
-                    if (attacker.hasKillingBlow === true) {
-                      return attacker._embedded.corporation.name;
-                    }
-                    return '';
-                  })
-                }
-                attackerCount={kill._embedded.attackers.length}
-                zone={kill._embedded.zone.name}
-                date={kill.date}
-              />
-            );
-          }),
-        });
-      });
-  }
-
-  componentDidMount() {
-    this.populateTableFromAPI(this.state.page);
-  }
-
-  render() {
-    const loaded = this.state.loaded;
-    var classes = "button is-large is-dark";
-    var imgClass = "is-invisible";
-    if (!loaded) {
-      imgClass = "";
-    }
-    return (
-      <div>
-        <Victims kills={this.state.kills} />
-        <div className="buttons">
-          <a disabled={!loaded} className={classes} onClick={this.handleClickFirst}>First</a>
-          {this.state.page > 1 && <a disabled={!loaded} className={classes} onClick={this.handleClickPrev}>Prev</a>}
-          {this.state.page < this.state.page_count && <a disabled={!loaded} className={classes} onClick={this.handleClickNext}>Next</a>}
-          <a disabled={!loaded} className={classes} onClick={this.handleClickLast}>Last</a>
-          <img className={imgClass} src={loadingGif} alt="loading..." />
-        </div>
-
-      </div>
-    );
-  }
-}
-
-class KillDetail extends Component {
-  state = {
-    'date': null,
-    'damageReceived': null,
-    'agent': {
-      'name': null,
-    },
-    'corporation': {
-      'name': null,
-    },
-    'robot': {
-      'name': null,
-    },
-    'zone': {
-      'id': null,
-    },
-    'attackers': [],
-  };
-
-  componentDidMount() {
-    fetch(new URL('/killboard/kill/' + this.props.match.params.killId, apiUrl), {
-      method: 'GET',
-      headers: new Headers({
-        "Accept": "application/*+json",
-      }),
-    })
-      .then(response => {
-        return response.json();
-      })
-      .then(json => {
-        this.setState({
-          'date': json.date,
-          'damageReceived': json.damageReceived,
-          'agent': json._embedded.agent,
-          'corporation': json._embedded.corporation,
-          'robot': json._embedded.robot,
-          'zone': json._embedded.zone,
-          'attackers': json._embedded.attackers.map(function (attacker) {
-            return (
-              <Attacker
-                key={attacker.id}
-                {...attacker}
-                totalDamageDealt={json.damageReceived}
-                agent={attacker._embedded.agent}
-                robot={attacker._embedded.robot}
-                corporation={attacker._embedded.corporation}
-                totalEcmAttempts={attacker.totalEcmAttempts}
-                energyDrain={attacker.energyDispersed}
-                sensorSurpressions={attacker.sensorSurpressions}
-              />
-            );
-          }),
-        });
-      });
-  }
-
-  render() {
-    return (
-      <div>
-        <div className="columns">
-          <div className="column">
-            <div className="has-text-centered" style={{ marginBottom: '2em' }}>
-              <div>
-                <p className="heading">Agent</p>
-                <p className="subtitle">{this.state.agent.name}</p>
-              </div>
-            </div>
-            <div className="has-text-centered" style={{ marginBottom: '2em' }}>
-              <div>
-                <p className="heading">Corporation</p>
-                <p className="subtitle">{this.state.corporation.name}</p>
-              </div>
-            </div>
-            <div className="has-text-centered" style={{ marginBottom: '2em' }}>
-              <div>
-                <p className="heading">Robot</p>
-                <p className="subtitle">{this.state.robot.name}</p>
-              </div>
-            </div>
-          </div>
-          <div className="column">
-            <div className="has-text-centered" style={{ marginBottom: '2em' }}>
-              <div>
-                <p className="heading">Zone</p>
-                <p className="subtitle">{this.state.zone.name}</p>
-              </div>
-            </div>
-            <div className="has-text-centered" style={{ marginBottom: '2em' }}>
-              <div>
-                <p className="heading">Date</p>
-                <p className="subtitle">{this.state.date}</p>
-              </div>
-            </div>
-            <div className="has-text-centered" style={{ marginBottom: '2em' }}>
-              <div>
-                <p className="heading">Damage received</p>
-                <p className="subtitle">{Math.round(this.state.damageReceived * 100) / 100}</p>
-              </div>
-            </div>
-          </div>
-        </div>
-        <div>
-          <h1 className="title is-4">
-            Attackers
-          </h1>
-          <div>
-            <Attackers attackers={this.state.attackers} />
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
-
-class Attackers extends Component {
-  render() {
-    return (
-      <div style={{ overflow: 'auto', overflowY: 'hidden' }}>
-        <table className="table is-striped is-fullwidth dark">
-          <thead>
-            <tr>
-              <th></th>
-              <th></th>
-              <th>Agent</th>
-              <th>Robot</th>
-              <th>Damage dealt</th>
-              <th>EWAR</th>
-              <th></th>
-            </tr>
-          </thead>
-          <tbody>
-            {this.props.attackers}
-          </tbody>
-        </table>
-      </div>
-    );
-  }
-}
-
-class Attacker extends Component {
-  render() {
-    return (
-      <tr id="row" className="attacker dark">
-        <td></td>
-        <td><img className="bot-icon" src={resolveIcon(this.props.robot.name)} alt="robot-icon" /></td>
-        <td><strong>{this.props.agent.name} {this.props.hasKillingBlow && <span className="tag is-danger"> Killing blow! </span>} </strong><br />{this.props.corporation.name}</td>
-        <td><strong>{this.props.robot.name}</strong><br /></td>
-        <td><strong>{Math.round(this.props.damageDealt * 100) / 100}</strong><br />{Math.round(this.props.damageDealt / this.props.totalDamageDealt * 10000) / 100}%</td>
-        <td>
-          {this.props.energyDrain > 0 && <span><strong>Energy Drained:</strong>  {Math.round(this.props.energyDrain)} <br /></span>}
-          {this.props.sensorSurpressions > 0 && <span><strong>Sensor Surpressions:</strong>  {this.props.sensorSurpressions} <br /></span>}
-          {this.props.totalEcmAttempts > 0 && <span><strong>ECM Attempts:</strong>  {this.props.totalEcmAttempts}<br /> </span>}
-        </td>
-        <td></td>
-      </tr>
-    );
-  }
-}
-
-class Victims extends Component {
-  render() {
-    return (
-      <div style={{ overflow: 'auto', overflowY: 'hidden' }}>
-        <table className="table is-striped is-fullwidth dark">
-          <thead>
-            <tr id="row">
-              <th></th>
-              <th></th>
-              <th>Robot</th>
-              <th>Victim</th>
-              <th>Killing blow</th>
-              <th>Location</th>
-              <th></th>
-              <th></th>
-            </tr>
-          </thead>
-          <tbody>
-            {this.props.kills}
-          </tbody>
-        </table>
-      </div>
-    );
-  }
-}
+import HeroSection from './components/HeroSection';
+import FooterSection from './components/FooterSection';
+import KillBoard from './components/KillBoard';
+import KillDetail from './components/KillDetail';
 
 class App extends Component {
   render() {
