+++ src/components/FooterSection.js
@@ -1,0 +1,18 @@
+import React, { Component } from "react";
+
+class FooterSection extends Component {
+    render() {
+      return (
+        <footer className="footer gray">
+          <div className="container">
+            <div className="content has-text-centered">
+              <p>
+                <strong>Open Perpetuum Killboard</strong> is part of Open Perpetuum, an open-source project. Support us on <a href="https://www.patreon.com/openperpetuum">Patreon</a>!
+              </p>
+            </div>
+          </div>
+        </footer>
+      );
+    }
+  }
+  export default FooterSection;
